pyproject.toml
[project]
name = "trader-web"
version = "0.1.0"
description = ""
authors = [
    {name = "RodResq",email = "rodrigoresque@yahoo.com.br"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.7,<6.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "django-debug-toolbar (>=5.1.0,<6.0.0)",
    "aiohttp (>=3.12.4,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "markdown (>=3.8.2,<4.0.0)",
    "django-filter (>=25.1,<26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
